Description:
Brings the logger port up to date with master_devel. Five conflicts, plus three
follow-on fixes the merge exposed that git could not have flagged.

- configure.ac and mtcr_ul/Makefile.am: both sides appended to the same lists
  (AC_OUTPUT, and the previously-empty libmtcr_ul_la_DEPENDENCIES). Kept both.

- mft_exceptions/Makefile.am: master_devel dropped lib_os_api from _LIBADD to
  stop its objects landing in two archives at once. Same reasoning applies to
  libmft_logger, which this branch only ever wanted for build ordering, so the
  merged rule keeps their _LIBADD and both entries in _DEPENDENCIES.

- efuse/Makefile.am: took their move of $(JSON_LIBS) out of _DEPENDENCIES (the
  FreeBSD link fix) and kept this branch's removal of liblogger.la, which became
  dead when efuse.cpp stopped using the legacy logger.

- efuse/efuse.cpp: master_devel extracted the fuse-query body into
  query_one_fuse(). Rather than hand-merge, their file was taken whole and the
  legacy-logger conversion re-applied over it, so all 13 of their call sites move
  to the EFUSE layer instead of the 10 that existed when this branch started.

Three things only surfaced at build time:

- mft_core/device/device_info is a new consumer of MftGeneralException.h, which
  now includes mft_logger.h, so it needs mft_core/mft_core_utils on its include
  path like the other four consumers.

- master_devel added a DBG_PRINTF call site in mtcr_ul_icmd_cif.c after this
  branch removed the macro. Converted the same way as the other 137.

- libmtcr_ul now links both libdevice_info and libmft_logger, and both embedded
  lib_os_api, so every OS API symbol was multiply defined. libmft_logger keeps
  lib_os_api in _DEPENDENCIES for ordering only; the hand-rolled libmft_logger.so
  rule names the archive explicitly since it has no other source for it. This is
  the same trap master_devel documented on mft_exceptions.

Tested OS: Linux (x86_64)
Tested devices: N/A (offline)
Tested flows: full build clean after the merge; libmft_logger.a confirmed to no
longer carry the operating_system_api objects; libmft_logger.so still exports
mft_log/mft_log_fmt with no undefined OS API symbols; layer gating re-checked on
COMMON and MFT_CORE (silent with no config, emitted at debug, and silent when
only the other layer is enabled).

Issue: 4887718

+++ mtcr_freebsd/mtcr_ul.c
@@ -3299,7 +3299,7 @@ int is_zombiefish_device(mfile* mf)
     }
 }
 
-void set_fwctl_dev(char* fwctl_dev, u_int16_t domain, u_int8_t bus, u_int8_t dev, u_int8_t func)
+void set_fwctl_dev(char* fwctl_dev, u_int32_t domain, u_int8_t bus, u_int8_t dev, u_int8_t func)
 {
     (void)fwctl_dev;
     (void)domain;
@@ -3308,7 +3308,7 @@ void set_fwctl_dev(char* fwctl_dev, u_int16_t domain, u_int8_t bus, u_int8_t dev
     (void)func;
 }
 
-void open_fwctl_dev(mfile* mf, u_int16_t domain, u_int8_t bus, u_int8_t dev, u_int8_t func)
+void open_fwctl_dev(mfile* mf, u_int32_t domain, u_int8_t bus, u_int8_t dev, u_int8_t func)
 {
     (void)mf;
     (void)domain;
